Split output package into multiple files (#31)
Split the monolith output.go (301 lines, 30+ types) into logical groups:
- printer.go: Printer struct, methods, and ErrorResult
- types_skill.go: skill-related result types
- types_platform.go: platform-related result types
- types_overlap.go: overlap/recommend result types

Consolidate structurally identical PlatformInstallEntry and
PlatformUninstallEntry into a single PlatformActionEntry type.
